+++ CHANGELOG.md
@@ -41,6 +41,7 @@ end
 ```
 
 - [@kddeisz] - Handle string literals that start with `%Q`.
+- [@kddeisz] - Handle question method methods in the predicate of an if with a comment in the body.
 
 # [0.22.0] - 2020-12-08